Porting the web interface of www.emergency.vic.gov.au to Android.
I've found that the mobile website tends to crash the android browser.
In addition, although there is a pretty nice looking mobile app
available in the play store (commisioned by the victorian government,
if I'm not mistaken) it is propriatery, and also requires Google
services to be installed so that push notifications can work. Even
if they do fix the issue with push notifications (i.e. allow those
of us without Google services installed to use it), I still feel like
this is the type of app that should be freely available for the
community to develop and participate with.